+++ include/entity.h
@@ -1067,6 +1067,15 @@ typedef struct {
     u16 unk8A;
 } ET_801B0AA4;
 
+typedef struct {
+    byte pad[4];
+    u16 unk80;
+    u16 unk82;
+    s16 unk84;
+    s16 unk86;
+    s16 unk88;
+} ET_Owl;
+
 typedef union { // offset=0x7C
     struct Primitive* prim;
     char stub[0x40];
@@ -1167,4 +1176,5 @@ typedef union { // offset=0x7C
     ET_LifeUpSpawn lifeUpSpawn;
     ET_AxeKnight axeknight;
     ET_801B0AA4 et_801B0AA4;
+    ET_Owl owl;
 } Ext;
